projects/wispr_flow_clone/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.23.2
python-multipart==0.0.7
openai==1.54.3
python-dotenv==1.0.1
